+++ xhpterm/conmgr.h
@@ -18,7 +18,7 @@ struct conmgr {
 
 struct conmgr * conmgr_connect (enum e_contype type, char *hostname, int port);
 void conmgr_read (struct conmgr *con);
-void conmgr_send (struct conmgr *con, char *buf, int nbuf);
+void conmgr_send (struct conmgr *con, char *buf, size_t nbuf);
 void conmgr_send_break (struct conmgr *con);
 void conmgr_close (struct conmgr *con);
 void conmgr_rxfunc (int32_t refcon, char *buf, size_t nbuf);
